Moved tests to test plugin and changed export settings of layered plugin.
