+++ plugins/org.eclipse.elk.alg.layered/META-INF/MANIFEST.MF
@@ -7,8 +7,7 @@ Bundle-Vendor: Eclipse Modeling Project
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: com.google.guava;bundle-version="15.0.0",
  org.eclipse.elk.graph,
- org.eclipse.elk.core,
- org.junit;bundle-version="4.12.0"
+ org.eclipse.elk.core
 Export-Package: org.eclipse.elk.alg.layered,
  org.eclipse.elk.alg.layered.compaction.components,
  org.eclipse.elk.alg.layered.compaction.oned,
