+++ test/org.eclipse.elk.alg.test/META-INF/MANIFEST.MF
@@ -10,3 +10,6 @@ Require-Bundle: com.google.guava;bundle-version="15.0.0",
  org.eclipse.elk.alg.layered,
  org.junit;bundle-version="4.12.0"
 Bundle-Vendor: Eclipse Modeling Project
+Export-Package: org.eclipse.elk.alg.test.layered.intermediate.greedyswitch,
+ org.eclipse.elk.alg.test.layered.p3order,
+ org.eclipse.elk.alg.test.layered.p3order.counting
