+++ test/org.eclipse.elk.alg.test/src/org/eclipse/elk/alg/test/layered/intermediate/greedyswitch/GreedySwitchProcessorTest.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Kiel University - initial API and implementation
  *******************************************************************************/
-package org.eclipse.elk.alg.layered.intermediate.greedyswitch;
+package org.eclipse.elk.alg.test.layered.intermediate.greedyswitch;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
