+++ test/org.eclipse.elk.alg.test/src/org/eclipse/elk/alg/test/layered/p3order/LayerSweepTypeDeciderTest.java
@@ -1,4 +1,4 @@
-package org.eclipse.elk.alg.layered.p3order;
+package org.eclipse.elk.alg.test.layered.p3order;
 
 import static org.junit.Assert.assertTrue;
 
@@ -8,10 +8,12 @@ import org.eclipse.elk.alg.layered.graph.LGraph;
 import org.eclipse.elk.alg.layered.graph.LNode;
 import org.eclipse.elk.alg.layered.graph.LPort;
 import org.eclipse.elk.alg.layered.graph.Layer;
-import org.eclipse.elk.alg.layered.intermediate.greedyswitch.TestGraphCreator;
+import org.eclipse.elk.alg.layered.p3order.GraphData;
+import org.eclipse.elk.alg.layered.p3order.LayerSweepTypeDecider;
 import org.eclipse.elk.alg.layered.p3order.LayerSweepCrossingMinimizer.CrossMinType;
 import org.eclipse.elk.alg.layered.p3order.counting.AbstractInitializer;
 import org.eclipse.elk.alg.layered.properties.LayeredOptions;
+import org.eclipse.elk.alg.test.layered.intermediate.greedyswitch.TestGraphCreator;
 import org.eclipse.elk.core.options.PortSide;
 import org.junit.Test;
