+++ test/org.eclipse.elk.alg.test/src/org/eclipse/elk/alg/test/layered/p3order/counting/BinaryIndexedTreeTest.java
@@ -8,11 +8,12 @@
  * Contributors:
  *     Kiel University - initial API and implementation
  *******************************************************************************/
-package org.eclipse.elk.alg.layered.p3order.counting;
+package org.eclipse.elk.alg.test.layered.p3order.counting;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
+import org.eclipse.elk.alg.layered.p3order.counting.BinaryIndexedTree;
 import org.junit.Test;
 
 // CHECKSTYLEOFF MagicNumber
